vit-sports-intelligence/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
SQLAlchemy[asyncio]==2.0.36
aiosqlite==0.20.0
asyncpg==0.30.0
alembic==1.14.0
pydantic==2.10.4
python-dotenv==1.0.1
python-multipart==0.0.20
httpx==0.28.1
tenacity==9.0.0
numpy==2.2.1
beautifulsoup4==4.12.3
lxml==5.3.0
celery==5.4.0
redis==5.2.1
scikit-learn==1.6.0
joblib==1.4.2
pandas==2.2.3
aiohttp==3.11.11
